+++ Examples/Example App/ExampleApp/App.swift
@@ -8,14 +8,7 @@ struct ExampleApp: App {
 	init() {
 		// Initialize the elvah Charge SDK as soon as possible in your app's lifecycle.
 		// This initializer is a good place to do this, but you can also use an `AppDelegate`.
-		Elvah
-			.initialize(
-				with: Elvah.Configuration(
-					apiKey: "YOUR_API_KEY",
-					theme: .default,
-					store: .standard
-				)
-			)
+		Elvah.initialize(with: .simulator)
 	}
 
 	var body: some Scene {
